# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"OrderDTO.Item": {
				"properties": {
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"price",
					"quantity",
					"productId"
				],
				"type": "object"
			},
			"OrderDTO.Order": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderDTO.Item"
						},
						"type": "array"
					},
					"creationDate": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					}
				},
				"required": [
					"items",
					"creationDate",
					"value",
					"orderId"
				],
				"type": "object"
			},
			"OrderDTO.CreateItem": {
				"properties": {
					"valorItem": {
						"type": "number",
						"format": "double"
					},
					"quantidadeItem": {
						"type": "number",
						"format": "double"
					},
					"idItem": {
						"type": "string"
					}
				},
				"required": [
					"valorItem",
					"quantidadeItem",
					"idItem"
				],
				"type": "object"
			},
			"OrderDTO.CreateOrder": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderDTO.CreateItem"
						},
						"type": "array"
					},
					"dataCriacao": {
						"type": "string"
					},
					"valorTotal": {
						"type": "number",
						"format": "double"
					},
					"numeroPedido": {
						"type": "string"
					}
				},
				"required": [
					"items",
					"dataCriacao",
					"valorTotal",
					"numeroPedido"
				],
				"type": "object"
			},
			"OrderDTO.ListOrders": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/OrderDTO.Order"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"totalPages",
					"data"
				],
				"type": "object"
			},
			"OrderDTO.UpdateOrder": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderDTO.CreateItem"
						},
						"type": "array"
					},
					"dataCriacao": {
						"type": "string"
					},
					"valorTotal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"dataCriacao",
					"valorTotal"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "jitterbit-test",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/order": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Pedido criado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderDTO.Order"
								}
							}
						}
					},
					"409": {
						"description": "Um ou mais identificadores de item já existe no banco"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderDTO.CreateOrder"
							}
						}
					}
				}
			}
		},
		"/order/list": {
			"get": {
				"operationId": "ListOrders",
				"responses": {
					"200": {
						"description": "Listagem concluida",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderDTO.ListOrders"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/order/{orderId}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Pedido encontrado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderDTO.Order"
								}
							}
						}
					},
					"404": {
						"description": "Pedido não encontrado"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"description": "Pedido deletado",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Pedido não encontrado"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Pedido editado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderDTO.Order"
								}
							}
						}
					},
					"404": {
						"description": "Pedido não encontrado"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderDTO.UpdateOrder"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}